osx: packagemaker: look into applications dir for packagemaker
In newer osx versions, the packagemaker might be installed manually
as it doesn't seem to come bundled with XCode anymore.

In this case, the usual location is for it to be installed at
the Applications directory. This patch makes the packager search
for it in the old directory and in Applications