recipes: workaround for recipes that build too fast
Some recipes just copy files around and are performed very fast. This
confuses the new files listing for universal builds as it relies on
the timestamps of the new files to select what are the new ones.

When the build is faster than the FS timestamp resolution, old files and
new files can be mixed together, causing old files to be removed from the
final prefix destination dir. A case of this is when gst-sdk-shell is built.

This patch prevents this by waiting a little before attempting to install files to
the prefix, so that the new files have a timestamp difference to the old ones