recipes/custom.py: Add support for SourceType.TARBALL in static plugin builds
custom.py was assuming that all recipes using it are of type GIT or GIT_TARBALL.
This is a hacky attempt (just as hacky as custom.py itself) to make it support
tarballs as well.

Note that cerbero already has really hacky support for recipe names that don't
match the tarball unpacked directory name, and that's required to make this work.