osx/ios: Don't include the arch path for non-universal builds
We build everything with libdir=lib, not lib/$arch. Things fail
if we move libraries around because Mach binaries have the complete
path to the library embedded instead of just the soname.

Also prevents us from creating some dead links...

There was probably a reason for this code originally, but nowadays
it just breaks. Leave it there but commented out.