packages: Build the list of non-devel files correctly
When no categories are specified for a recipe, the list of non-devel files only
excludes the variable files_devel. Instead, it should exclude all variables
ending in _devel to mirror the list of devel files which includes all variables
that end with _devel. Otherwise, the non-devel package will contain devel files.