spandsp: fix dllexport/dllimport usage on windows
LIBSPANDSP_EXPORTS must only be defined while building libspandsp,
otherwise code that uses spandsp symbols will be using them with
the dllexport attribute and the binaries will probably fail at runtime.

The change in config_sh is required because we patch the Makefile.am,
so automake needs to run again, and while "make" will detect that,
the code in the tarball is configured to run "automake-1.13", which
will fail on most systems nowadays. Calling autogen.sh again ensures
that we call the correct automake.