gstreamer-1.0.mk: fix missing ";" bug
Several command invocations wrapped in for-each are missing ";"
characters, which meant they get expanded on one line instead of
separate lines. For some for-each invocations, you can get away with
this and it will be OK, but for something like host-cp, it changes the
meaning. As a result, some host-cp invocations will fail with errors
like:

cp: target ‘X’ is not a directory

Fix this by adding the necessary semicolons.

https://bugzilla.gnome.org/show_bug.cgi?id=759380