cerbero: Allow alternate arch configs in universal builds
This allows us to specify universal_archs as a map of arch -> config
file (which can be None). If specified, the config file provides
overrides to the default config for that architecture.

This is useful on Android, for example, where we want to specify
DistroVersion as Lollipop+ for ARM64 and x86-64, but use a lower version
for the other arches.

https://bugzilla.gnome.org/show_bug.cgi?id=770631