cerbero: Try to redownload corrupted tarballs
Earlier, when we encountered an incomplete or corrupted tarball, we
would error out with a CRC check error (for tarballs) or similar. This
error would not fix itself if you build the recipe from scratch because
the 'fetch' stage just checks if the file exists, and if it does,
assumes that it downloaded successfully.

You had to manually remove the file, and then when 'extract' errored
out because the file didn't exist, you'd have to select '1' to build the
recipe from scratch. This is extremely confusing.

Now we just overwrite and try to redownload the file once if it's
corrupted or partially downloaded.