Workaround broken XCode 8 SDK weak symbol detection
macOS 10.12 added a bunch of new symbols: clock_gettime, getentropy,
mkostemp, etc. These are available in the macOS 10.12 SDK. However,
XCode 8 now only ships that version of the SDK, and because of the way
the symbols are defined, they are always detected as being available
regardless of what OS X version you are actually targetting.

Long story short, the only way around this is to workaround it by
forcibly disabling detection of these symbols when we're targetting OS
X versions older than 10.12. For more details, see:
https://bugzilla.gnome.org/show_bug.cgi?id=772451