cerbero: Download from our tarball mirror on failure
This covers the case when the server is down or returns an error code,
but not if the server is stupid enough to return a webpage with an HTML
message. For that we need to add a file size and sha256sum check.

Once we have that, we can even add support for continuing downloads.

TODO: Fix tarball mirror URL