gst-libav: Use gas-preprocessor for assembly on iOS
Make the assembler for libav be built via gas-preprocessor
on iOS, as recommended by libav. The latest arm64 assembly
breaks in clang's internal assembler due to some bug
handling macro arguments, and this fixes it.

Add a patch reverting a commit in gst-libav so our
assembler setting is actually passed through to libav
instead of being overridden to be $CC