gst-libav: Let gst-libav install the static deps
In this patch we remove the unused libswscale library. We also
no longer install the static deps, this is now done by gst-libav.
This was needed to make static builds usable without cerbero. We
still produce .la file as these are requires by our linker script
for Android ("libtool ... -static-libtool-libs" works without that
as we properly include the deps and order of deps in the plugin .la
file).