config/android: fix include/lib paths for non-universal builds
The directory layout is slightly different for universal vs
non-universal builds and the there's currently no way to be able to
detect that when parsing the configuration.

Add such a way by adding the 'universal' string to the variants when
parsing the children of a universal configuration.

While we're here also document some paths used in the android build and
the sysroot/isysroot difference and why that's needed with NDK's Unified
headers.

This fixes the glib's failure to find the correct iconv.h header/library
now that the NDK ships it's iconv.h with all functions #if-ed out and
unusable.

https://ci.gstreamer.net/job/cerbero-android/8220/console