cerbero: Add a urllib2 implementation for downloads
The wget that ships with msys is too old and cannot use TLSv1.2 or
newer, which GitHub started requiring a few months ago:

https://githubengineering.com/crypto-removal-notice/

The implementation is very basic right now and is only used on Windows.
For instance, it does not do certificate checking, but that was always
broken on Windows anyway.