cerbero: Use dlltool to find dlls instead of guessing
On Windows, DLLs can literally be called anything.dll, so any regex we
use will fail eventually. In this case, it fails on openssl because
the output DLLs are called libcrypt-1_1-x64.dll and libssl-1_1-x64.dll

Instead, use dlltool -I on the import library to find the real dll.