openssl.recipe: Fix build on macos and ios
CFLAGS cannot be passed to Configure because it detects -arch x86_64
as a target directive and errors out. We also can't rely on CFLAG
passing it everywhere, so we have to set it with CC.

This also means we don't need the patch that adds support for
detecting -isysroot, etc, since those should always be passed
correctly now.