Fix generation of fat shared libraries on macOS
We have to properly check which of the prefixes is a parent directory of
our file, not via string comparisions. Otherwise "/foo/x86" and
"/foo/x86_64" or both considered parents of "/foo/x86_64/bar" and which
one is stripped from the path depends on the order.

Thanks to Mathieu Duponchelle for the non-ugly generator-based function
for doing this check.