source: Fix LD_LIBRARY_PATH when handling git commands
For non-cross-compilation use-cases, we might end up building libraries
in cerbero that should *not* be used by host commands.

The case in point was the host "git" command picking up a mix of system
libraries and of libraries that we built in cerbero.

To avoid that, before running git commands switch to an environment
which uses the "host" LD_LIBRARY_PATH environment

https://bugzilla.gnome.org/show_bug.cgi?id=794830