Port glib and glib-networking recipes to meson
Also, always use gnutls system trust for glib-networking. This is
backported from latest master, and we should fix our gnutls config to
work with this. Porting it in early so we have some time to deal with
this before the next glib-networking stable release.

Tested on fedora linux, cross-win32/64, cross-android-universal, darwin,
cross-ios-universal, and windows/mingw.