glib.recipe: Use autotools value for G_MODULE_SUFFIX
This was changed in Meson, but it's a bad idea because then we can't
find plugins built with Autotools.

Upstream bug: https://gitlab.gnome.org/GNOME/glib/issues/1413

Also remove a useless post_install from gstreamer-1.0.recipe with
the nice side-effect of rebuilding it which is needed for this to
work.