cerbero: Fix setting of env vars in meson recipes
Unsetting *FLAGS using self.new_env in __init__ means the *FLAGS set
in the recipe are completely ignored because they are first set, then
unset.

Instead, unset them in configure() and then re-add the values that
were set in the recipe using self.append_env.

This really needs to be re-architectured at some point, but it
requires a bunch of rework in config.py