gst-libav-1.0.recipe: Fix license configure option
The option is --disable-gpl not --enable-lgpl, which means we didn't
pass the option down to ffmpeg.

Thankfully, both gst-libav and ffmpeg default to --disable-gpl, so this
did not cause any isues. Verified by checking ffmpeg's configure output:

    Configuring included Libav instance with args

    ...

    License: LGPL version 2.1 or later

The default for ffmpeg is also --disable-nonfree + --disable-version3
so the license specified in the recipe is correct.

https://bugzilla.gnome.org/show_bug.cgi?id=797041