cerbero: Properly fetch the value of PATH from MSVC
We were accidentally using the full value of PATH set in the VS command
prompt as the PATH value that contains Visual Studio paths. Only
a subset of that value contains Visual Studio paths.

As a part of this, now the rest of the code prepends PATH fetched from
VS to the existing PATH value and uses the other env vars as-is by
overwriting.