cerbero/bootstrap: Set and unset the build-tools env properly
When we cook the build-tools recipe, we temporarily override the
current environment with one derived from the build-tools
configuration, cook, and then re-set back to the old env.

We stopped doing this correctly after splitting out the 'fetch' step.
This caused a bootstrap failure because part of our build step was
using the wrong configuration.