cerbero: Store local sources in a user-wide location
Since we store local sources inside the cerbero home dir (build/),
we download multiple copies of the sources for each cerbero clone or
if the user deletes the homedir or moves the git repository.

We still keep the old value for non-default cases where the user has
a configuration that overrides the default value of home_dir.

https://bugzilla.gnome.org/show_bug.cgi?id=797323