build/recipe: fix .la generation with project inter-dependencies
e.g. libgstbase-1.0 depends on libgstreamer-1.0 however due to sorting
libgstbase-1.0.la would always be generated before libgstreamer-1.0 and
not include a reference to the .la file but a -lgstreamer-1.0 line.

The gst android .la-parser-in-makefile simply passes through -l
arguments and would result in linking the shared library instead of the
static library as required.