libunwind.recipe: Run autoreconf to update aclocal
We copy ltmain.sh for all autotools recipes now, and run autoreconf
for some of them because of that. This recipe was missed.

libtool: Version mismatch error.  This is libtool 2.4.6, but the
libtool: definition of this LT_INIT comes from libtool 2.4.2.
libtool: You should recreate aclocal.m4 with macros from libtool 2.4.6
libtool: and run autoconf again.