cookbook: Allow recipes relocation
As of now, if a recipe had changed location we would always assume a
rebuild was needed. During the development of the new Gitlab CI, it came
a need to reuse cerbero caches from a different build directory.

To make the relacation safe, we simply force doing hash comparision
whenever we detect that the file path have changed.