Add manifest support
The build manifest is used in CI to fixate the Git revision of certain
project so that all builds of a pipeline are on the same reference. It
can also be used in order to re-produce a specific build.

To set a manifest, you can set manifest = 'my_manifest.xml" in your
configuration file, or use --manifest command line option. The command
line option will have precendence on any configuration file.