cerbero: Fix cross-ios framework merging inside SSH
The encoding when logging into a mac machine with SSH is ASCII, and
calling `file -bh` on some of our translation files now outputs
unicode information, in the script that the language for that
translation file uses. This leads to a stdout decode error.

We don't actually care about that since we only use the output for
categorising file types, so ignore it.

Recipe 'glib' failed at the build step 'merge'
Traceback (most recent call last):
  File "./cerbero/build/oven.py", line 155, in _cook_recipe
    stepfunc()
  File "./cerbero/build/recipe.py", line 559, in merge
    generator.merge_files(inputs, dirs)
  File "./cerbero/tools/osxuniversalgenerator.py", line 91, in merge_files
    self.do_merge(f, dirs)
  File "./cerbero/tools/osxuniversalgenerator.py", line 156, in do_merge
    action = self._detect_merge_action(full_filepaths)
  File "./cerbero/tools/osxuniversalgenerator.py", line 130, in _detect_merge_action
    ftype = self.get_file_type(f)
  File "./cerbero/tools/osxuniversalgenerator.py", line 122, in get_file_type
    return self._call(cmd)[0:-1] #remove trailing \n
  File "./cerbero/tools/osxuniversalgenerator.py", line 236, in _call
    output = output.decode(sys.stdout.encoding)
UnicodeDecodeError: 'ascii' codec can't decode byte 0xc3 in position 141: ordinal not in range(128)