universal: Improve universal sub-config loading
This is a partial fix, basically the path for universal sub-config was
derived from the first configuration file. In this patch, we now try to
derive from the first configuration file that has 'universal' in it's
name, and then fallback to to first configuration file path.

To reproduce this the issue, simply run:
  touch local.cbc
  ./cerbero-uninstalled -c local.cbc -c config/cross-android-universal.cbc show-config

Which fails with:
  Configuration Error: Configuration file $HOME/.cerbero/cross-android-armv7.cbc.cbc doesn't exist