cerbero: Fix wget errors while running fetch-bootstrap
wget: [...]/dist/linux_x86_64/lib/libgnutls.so.30: version `GNUTLS_3_6_3' not found (required by wget)

We need config._pre_environ to be the env that was set when Cerbero
was called. But while bootstrapping build tools, we instantiate that
again without resetting the env to that value, so it picks up the
current Cerbero env and sets it as the _pre_environ.

Clear the env and set it to the original env before instantiating
Config()