gst recipes: Disable plugins when building with MSVC
Some libraries have issues that cause us to be unable to include their
headers or link to their libraries when building with MSVC. Disable
them when we're building with MSVC on Windows.

With this, you can produce a mostly-complete gstreamer-1.0 package
that is built with MSVC on Windows.

The status of these plugins is tracked at:
https://gitlab.freedesktop.org/gstreamer/cerbero/issues/121