libvpx.recipe: Fix build with VS 2015
This reverts 35549ca9b8446e83dc739f7d30be19d58510b9e7

I was completely wrong in that commit while trying to fix
https://gitlab.freedesktop.org/gstreamer/cerbero/issues/90

The platform toolset is not the SDK, it's the tools being used to
build everything (msbuild, cl, etc). I verified that vpx continues to
build even if the only SDK installed is the Windows 10 SDK.

We can revisit this if someone reports an issue again.