cerbero: Use variant for export MSVC shell
We were using a separate winXX-msvc-shell.cbc configuration file
to signal to the main configuration file that we want to export for
an MSVC shell on Windows.

This meant that people had to first know that they were supposed to
use this, and then know that they could chain it with other
configuration files to get a shell with, f.ex., a different prefix.

We already have a mechanism for picking msvc vs mingw shell: the
variant. So now we use that.