cerbero: Use a different prefix and cache for MSVC
When the `visualstudio` variant is enabled, we will use a separate
prefix for dist, sources, logs, and for the cache itself. This avoids
user error while reusing the cerbero prefix.

It will also help with enabling Windows CI on GitLab because the build
cache must not be shared between MinGW and MSVC.