gensdkshell: Copy input env dictionary
Dictionaries are mutable, callers don't expect their env to be modified.
On top of this, the default dictionary is instantiated once which will
have side effect on the second call.

This is noticeable when looking at the gst-env generate script. All the
env values are duplicated (duplicated path, duplicated flags, etc).