cerbero/msvc: Better way to export the MSVC shell
Instead of exporting exactly once, check if the variable already has
the values we want to add, and don't add it if it does. This fixes
a potential issue where we set the some env var earlier and want to
append to it when exporting for MSVC, which would result in the var
not being set at all.