cerbero: Simplify async_call and async_check_call
We do not need all the kwargs that the sync versions have, and we can
simplify the implementation of the functions this way.

Now those two take either a list of strings, which will be run
directly without a shell, or a single string which will be passed to
`sh -c` because it probably as shell syntax like `&&` or `;`, or sets
env vars with shell syntax.

This is needed to ensure that MSYS scripts continue to work, and that
the MSYS binaries are selected when, f.ex., `find` is run.

Also rename async_check_call to async_call_output which is a more
accurate name.