cerbero/windows: Hack-fix for shell.async_check_call
On Windows, create_subprocess_exec with a PIPE fails while creating
a named pipe using tempfile.mktemp because we override os.path.join to
use / on Windows. Override the tempfile module's reference to the
original implementation.

I tried to get rid of the os.path.join hack, but it breaks
*everything*, and it's too late in the release cycle to play
whack-a-mole and convert everything to Path.as_posix().