cerbero/windwos: Fix searching of vpx dll on MSVC
vpx.lib is not an import library, but an MSVC static library. Running
dlltool against it outputs a lot of errors on stderr, and we do not
want to return that as the dllpath.

In general, we do not want errors in the output, and we also want the
command to actually fail when it errors out.