recipes: gst-validate: disable debug-viewer
We don't need it, and it causes errors on install due
to lack of permissions because the new meson python
module returns an absolute path into the python install
prefix for

  import('python').find_installation().get_path('purelib')

where the old deprecated python3 module returned a path
without a prefix. This is apparently on purpose, so nothing
for us to fix here.