Revert "cerbero: Set os.environ instead of editing it in-place"
This reverts commit d28f934ac6d744d2d6b444709c342ee66250aeff.

Otherwise subprocess invocations will not have the new environment,
which causes a CI failure on macOS during the install phase of this
recipe:

[(72/79) gstreamer-1.0-osx-framework -> install ]
/bin/sh: pkg-config: command not found
/bin/sh: pkg-config: command not found

We can get rid of os.environ manipulation completely in the future.
We already pass the `env` explicitly to most places, but there is
a chance of breakage, f.ex. code that uses `cerbero.ide.pkgconfig`.
So we can only do this in 1.17.x.