cerbero: Fix build on the Jenkins CI
Jenkins doesn't use --manifest or the `manifest` property. It uses the
`recipes_commits` property to set the commits to be used by Cerbero,
so we should also use that to decide when to disable usage of tarballs
for gstreamer recipes in custom.py

Also, the whole setup was broken for --manifest too because I didn't
test the initial implementation properly.