packages: Fix rpm generation
The root problem was that the --transform tar option would by default
apply the prefix on regular files AND links. This would result in the
links targetting packagename-version/libfoo.so instead of just libfoo.so

This would also exhibit another issue with tar usage, which is that when
you provide it with several times the same file, it would make hardlinks
of those duplicate files ... which would result in symbolic links
becoming hard links. Instead of that, ensure that the file list we
provide to tar is unique.

See #154