Add output_dir parameter to post_package()
When building a package, the -o option allows to specify an output
directory but this information is not passed to the post_package()
method.

If you need to perform post-package operations like uploading packages
files to a remote server or building a NuGet package from previous
files, you need to have access to the output folder location.

Adding the new output_dir parameter to the post_package() method, you
can now do whatever you need with freshly new created packages.