git: simplify the reset of the source branch.
Doing git reset --hard twice when combined with arbitrary cache
of the source dir in the CI can lead to a state with untracked
files and git refusing to checkout the cerbero-build branch.

Instead we can directly checkout the commit we want and reset
the build branch with a single command.