cerbero: Use the new subprocess API for unpacking
This avoids quoting issues. We should never, ever, use the old call
API for new code. A list of arguments is always going to be correct
and devoid of quoting issues, at least on platforms that natively
support argv[].

Also, pass -C as the first argument to tar because the order can
matter sometimes.