static: force rebuild of reverse deps for static recipes
When a recipe that creates a static library is rebuilt,
all reverse dependencies should be rebuilt to pick the last changes.
For example, if libvpx is configured to generate a static library
and the recipe is rebuilt, gst-plugins-bad should all be rebuilt
so that the plugin is linked against the latest build of libvpx.