build: change environment variables immediately inside __init__()
Otherwise various recipe.get_env() calls inside a recipe's prepare() will
not retrieve the correct value.  Fixes bitcode bundling in at least rtmp
and the MAKEFLAGS/CONFIG_SITE overriding.

We also need to pass the environment to be used earlier (on creation) for
this to work.

Contains an environment hack to not add PKG_CONFIG_PATH from the system
when compiling so that gobject-introspection does not pick up the
incorrect libraries.