windows: Search for the executable before running
On Windows, we can't pass the PATH variable through the env= kwarg to
subprocess.* and expect it to use that value to search for the
command, because Python uses CreateProcess directly.

Unlike execvpe, CreateProcess does not use the PATH env var in the env
supplied to search for the executable. Hence, we need to search for it
manually with shutil.which(). The alternative is to use shell=True,
which we don't like.