Fix async tasks on Windows
asyncio.Semaphore classes set their working event loop internally on
creation, so we need to ensure the proper loop has already been set by then.
This is especially important if we create global semaphores that are
initialized at the very beginning, since on Windows, the default
SelectorEventLoop is not available.