cerbero: Disable fancy build status output on MSYS
The default MSYS shell window doesn't handle ANSI escape sequences
correctly, which makes our fancy build status output which rewrites
the previous line completely unusable.

Not sure if this is fixable at all, but the perf improvements from
parallel recipe building are very important on Windows, so here we
are.