cerbero: Increase parallel builds on Windows
We run make with -j1 on Windows due to make bugs, so we can run more
build steps for recipes in parallel without contention.

However, we cannot use `self.jobs` as the build job limit because a lot
of CPU is used by the Windows Defender anti-virus which tracks the
work done by build system processes even if you add the cerbero build
directory to the exclusion list.