Only the dev pkg should contain symlinks for shared libraries w/o version number
Currently the main packages generetad with cerbero contain symlinks for
shared libraries without version number. For instance, executing:

```
./cerbero-uninstalled package base-system-1.0
```

The `base-system-1.0_1.16.1-1_amd64.deb` contains symlinks like
`libturbojpeg.so`, `libxml2.so` or `libz.so`.

These symlinks are also in the development package. And there is an error when the development
package is installed because the system tries to overwrite a file, which was installed by
the main package.

From debian site:
https://www.debian.org/doc/manuals/maint-guide/advanced.en.html

> Please note that the development package should contain a symlink for the
> associated shared library without a version number.
> E.g.: /usr/lib/x86_64-linux-gnu/libfoo.so -> libfoo.so.1

From the fedora site:
https://fedoraproject.org/wiki/Archive:BuildingPackagesGuide?rd=Docs/Drafts/BuildingPackagesGuide

> All headers, static libraries, libtool archives, *.so files, autotools,
> and pkgconfig files go in the -devel subpackage.

With a git bisect, looks like ba26e9c0 is the first bad commit.