cerbero-uninstalled: Fix running from cmd.exe
You can run cerbero from cmd.exe like so:

```
> C:\MinGW\msys\1.0\bin\sh.exe --login -c 'c:/path/to/cerbero-uninstalled [args]'
```

In this case, `dirname` returns a native-win32 path, not an MSYS unix
path. We should short-circuit and return that directly in that case.

PS: We cannot run cerbero directly from cmd.exe as a python script
because we need the MSYS environment for building Autotools recipes.