cerbero: Skip recipe errors when running a shell
We don't care about these because we often run with an env that
recipes can't be parsed in. F.ex., `openssl.recipe` errors out with
a `KeyError` when running `-v visualstudio shell` because we don't set
`CC` in that env.

The errors were being ignored anyway, but we shouldn't print them at
all.