cerbero: Don't redownload on *all* unpack failures
Sometimes zipfile.extractall returns `PermissionError`, such as when
the code that writes out permissions on extracted files gets
interrupted and some file being extracted gets written out with the
default permissions which do not include +w, so re-running the extract
can fail while trying to write to that same file.

This is obviously not a problem with the downloaded zip file. However,
that exception is a subclass of `IOError` so we think the zip file is
corrupted and download it again. The handling was similarly broken for
extracting tarballs with `tar` and the tarfile module too.

Now we try to be more specific about which errors are due to the file
being (possibly) corrupted.