cerbero/config: Avoid using env vars for compiler paths
If we pass only some compiler system include paths using env vars, the
order will be incorrect, and Meson will remove any `-isystem` args we
add ourselves to force a particular order if they're also in the env
var (which adds it to the default compiler search path). This will
break glib's search for iconv.h on cross-android.

If we always pass the correct list of system include directories using
the env var, cross compilation breaks because some autotools recipes
call the native (non-cross) compiler with the same env, and it will
pick up the wrong headers causing a build failure; f.ex., gmp on
cross-android.

The only solution is to never pass these args when cross compiling,
and to always pass it using the CFLAGS, CPPFLAGS, etc env vars.