recipes: Move from gnutls + openssl to only openssl
Previously, all our recipes were linking to either openssl or gnutls.
As a result, we were shipping *two* TLS implementations, which is
messy because it means consumers of Cerbero and our binaries had to
configure the certdb twice (thrice if you count glib-networking), do
a security audit of two TLS implementations, and generally had
increased binary sizes because of this.

Since we require openssl for the DTLS plugin (used by webrtcbin), and
all the other recipes can pick either openssl or gnutls, we have to go
with openssl.

This involved importing patches from vcpkg to port librtmp to the
latest OpenSSL 1.1.1 version since upstream is non-existent.

We have also deleted the following recipes: gmp, nettle, libtasn1,
gnutls.