cerbero: Only add system libs on linux, when not cross-compiling
This feature is specifically only for libraries with pkg-config files
that are shipped on all Linux distros and are:

1. Linux-specific libraries or features
2. Also available on other platforms (for which we ship a recipe), but
   we don't want to use that recipe on Linux for various reasons

For example, it can cause issues with `LD_LIBRARY_PATH` and build
tools provided by the system on Linux.

So, just make it a no-op on other platforms instead of setting it
conditionally in recipes. We never want it to be set on other
platforms.