utils: handle nested run_tasks()
As used when building universal variants (ios, android)

We need to propagate the CancelledError from the sub-run_tasks() to the
parent run_tasks() so that the parent knows to stop rather than
continuing.

We attempt to avoid raising the CancelledError if another exception
is thrown somewhere else as a task being cancelled is not an
interesting result from a failure.