utils: fix spammy BlockingIOError on older python versions
Python < 3.8 used SIGCHLD to reap child proceses and would poll every
process on every SIGCHLD (asyncio.SafeChildWatcher).  This may become
overwhelming with lots of short-lived processes.  We don't actually
need this functionality of polling individual processes so we can use
the asyncio.FastChildWatcher instead.

We keep the default asyncio.ThreadedChildWatcher for python 3.8 as it
does not produce BlockingIOError