cerbero/windows: Don't try to use objc/objc++ compilers
Meson 0.53 can correctly detect MinGW objc/objc++ compilers, which
means that if you follow our instructions and install MSYS/MinGW, you
will have those available, which will be found by Meson.

This has the unintended side-effect that Meson will then use the
objc++ compiler to find libraries that it's searching for in
`dependency()` calls, since it has higher priority than the C++
compiler (which is MSVC++ in our case).

So Meson will end up preferring `libfoo.dll.a` over `foo.lib` when
looking for import libraries. Unfortunately, some of those are
malformed (such as `libssl.dll.a` and `libcrypto.dll.a`) and the DLLs
will fail to load some symbols at runtime, such as `DTLS_method`.

Fixes https://gitlab.freedesktop.org/gstreamer/cerbero/issues/246

To fix the duplication in the outer env and in the MSVC env, we will
need to completely switch over to using `mingw_toolchain_env` for the
MinGW/GCC environment so that we don't need to wipe the outer env when
building MSVC recipes.