cache: Re-implememt cache on top of our external storage
The goal is to reduce egress on FDO server. In this patch, the cache is split
from fetch.py into it's own set of commands in cache.py. It now implements
fetch-cache, gen-cache and upload-chache. upload-cache can only work if the
SSH private key needed to upload has been set into CERBERO_PRIVATE_SSH_KEY
environment or is available in your .ssh folder. The key will be made availabe
through configuration in gstreamer/cerbero CI configuration and only
available to protected branches on the gstreamer namespace.

Backported-by: Nirbheek Chauhan <nirbheek@centricular.com>