pkg-config.recipe: Always auto-detect the prefix
With this, the `prefix` variable in .pc files will be automatically
detected based on the location of the .pc file. This allows for
portable prefixes.

This is the recommended configuration on Windows and macOS. We need to
enable this since we ship pkg-config with the release binaries now.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/634>