Add a new variant type for mappings for Visual Studio CRTs
You can now specify variants that are key-value pairs by specifying
`variant=value` as the variant, f.ex., the first mapping-variant is
`vscrt`:

$ cerbero-uninstalled -v vscrt=mdd package gstreamer-1.0

`md` is the default, aka release CRT, and `mdd` is the debug CRT.

The main advantage of this is to be able to use the debug CRT and also
build with optimizations enabled, since the only other way to link to
the debug CRT is by setting `--buildtype=debug`, which disables
optimizations.

This is useful when debugging, for example, Qt apps since it allows
you to link to the debug Qt libraries without making your gstreamer
code unusably slow by disabling all optimizations.

Note that code built with MinGW will still always use the release CRT,
but since this is for debugging, that should be fine.